Separate code from main to MatthewPage
